App-5-Flatmates-Bill-Web-App/requirements.txt
certifi
chardet
click
filestack-python
Flask
fpdf
idna
itsdangerous
Jinja2
MarkupSafe
requests
trafaret
urllib3
Werkzeug
WTForms
python-dotenv
